% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsBackendHdf5Peaks.R
\name{MsBackendHdf5Peaks-stash}
\alias{MsBackendHdf5Peaks-stash}
\alias{saveMsObject,MsBackendHdf5Peaks,PlainTextParam-method}
\alias{readMsObject,MsBackendHdf5Peaks,PlainTextParam-method}
\alias{saveObject,MsBackendHdf5Peaks-method}
\alias{saveMsObject,MsBackendHdf5Peaks,AlabasterParam-method}
\alias{readMsObject,MsBackendHdf5Peaks,AlabasterParam-method}
\title{MsBackendHdf5Peaks Stash}
\usage{
\S4method{saveMsObject}{MsBackendHdf5Peaks,PlainTextParam}(object, param, consolidate = FALSE)

\S4method{readMsObject}{MsBackendHdf5Peaks,PlainTextParam}(object, param, spectraPath = character())

\S4method{saveObject}{MsBackendHdf5Peaks}(x, path, consolidate = FALSE, ...)

\S4method{saveMsObject}{MsBackendHdf5Peaks,AlabasterParam}(object, param, consolidate = FALSE)

\S4method{readMsObject}{MsBackendHdf5Peaks,AlabasterParam}(object, param, spectraPath = character())
}
\arguments{
\item{object}{An \code{MsBackendHdf5Peaks} object.}

\item{param}{Either a \code{PlainTextParam} or \code{AlabasterParam}.}

\item{consolidate}{\code{logical(1)} whether in addition to the spectra metadata
also the peaks data file (in HDF5 format) should be stored in the stash
folder. Default is \code{consolidate = FALSE}.}

\item{spectraPath}{For \code{readMsObject()}: optional \code{character(1)} with the
path to the peaks data in HDF5-format (in case they are on longer
available in the folder referred to by the original stashed
\code{MsBackendHdf5Peaks} object).}

\item{x}{An \code{MsBackendHdf5Peaks} object.}

\item{path}{For \code{saveObject()}: \code{character(1)} with the path where the
object should be stored in.}

\item{...}{Currently ignored.}
}
\value{
\code{readMsObject()} returns an \link[Spectra:MsBackendHdf5Peaks]{Spectra::MsBackendHdf5Peaks}` object.
}
\description{
\code{MsBackendHdf5Peaks} classes can be stashed to (or read from) plain text
file-based or \emph{alabaster}-based formats using the \code{\link[MsStash:saveMsObject]{saveMsObject()}} and
\code{\link[MsStash:readMsObject]{readMsObject()}} functions combined with the \link[MsStash:PlainTextParam]{PlainTextParam} and
\link[MsStash:AlabasterParam]{AlabasterParam} parameter objects, respectively. In both cases, data files
are stored into the specified directory. By default, only the backend's
spectra metadata is stored in that folder. Setting parameter
\code{consolidate = TRUE} will also copy the HDF5-format peaks data files
(containing the \emph{m/z} and intensity values) of the backend into the folder
generating a self-consistent stash.

Details on the stored files are provided in the sections below.
}
\details{
A \code{MsBackendHdf5Peaks} stash will by default only contain the spectra
metadata (spectra variables) but no \emph{m/z} and intensity values. The reference
to the original HDF5-format peaks data files are stored as spectra variable
\emph{dataStorage}. If these files were moved or if the stash was copied to a
different computer, the path to these original data files needs to be
provided to the \code{readMsObject()} function with parameter \code{spectraPath}.
Alternatively, with \code{consolidate = TRUE}, it is also possible to \strong{copy} the
peaks data files to the stash directory generating a self-contained data
stash. Note however that in this case two copies of all data files exist (in
the original location \strong{and} the stash directory).
}
\section{Text-file format, \code{PlainTextParam}}{


The \code{saveMsObject()} function with the \code{PlainTextParam} stores the spectra
metadata (spectra variables) of an \code{MsBackendHdf5Peaks} to a plain tabulator
delimited text file with the name \emph{ms_backend_spectra_data.txt} in the
directory specified with parameter \code{path} of the \code{PlainTextParam} object.
Depending on parameter \code{consolidate} also the peaks data files (in HDF5
format) will be copied to the stash folder (with \code{consolidate = TRUE}).
}

\section{\emph{abalbaster}-based format, \code{AlabasterParam}}{


With \code{AlabasterParam}, the spectra metadata will be exported (or imported)
through the \emph{alabaster} framework. Similar to the \code{PlainTextParam},
\code{consolidate = TRUE} will also copy the HDF5-format peaks data files to the
stash directory.
}

\examples{

library(Spectra)
library(StashSpectra)
## Create an example MsBackendHdf5Peaks backend from a single mzML file
library(MsDataHub)
tmp <- backendInitialize(MsBackendMzR(), X20171016_POOL_POS_1_105.134.mzML())
be_h5 <- backendInitialize(MsBackendHdf5Peaks(), data = spectraData(tmp),
    hdf5path = file.path(tempdir(), "h5_backend"))
be_h5

d <- file.path(tempdir(), "example_hdf5")
ptp <- PlainTextParam(path = d)

## Store the object into the stash, including the peaks data files.
saveMsObject(be_h5, ptp, consolidate = TRUE)

## List the content of the folder: ms_backend_spectra_data.txt file
## with the spectra metadata and an HDF5 file with the peaks data:
dir(d)

## Restore the stashed object
res <- readMsObject(MsBackendHdf5Peaks(), ptp)

## Store the object in an alabaster-format stash
d <- file.path(tempdir(), "example_hdf5_2")

ap <- AlabasterParam(d)
saveMsObject(be_h5, ap)

## Check the content of the stash; with the default (`consolidate = FALSE`)
## no HDF5 data file was moved.
dir(d)

## Restore the object again
res <- readMsObject(MsBackendHdf5Peaks(), ap)
res
}
\author{
Johannes Rainer
}
