% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra.R
\name{SpectraStash}
\alias{SpectraStash}
\alias{saveMsObject,Spectra,PlainTextParam-method}
\alias{readMsObject,Spectra,PlainTextParam-method}
\alias{saveObject,Spectra-method}
\alias{saveMsObject,Spectra,AlabasterParam-method}
\alias{readMsObject,Spectra,AlabasterParam-method}
\title{Spectra Stash}
\usage{
\S4method{saveMsObject}{Spectra,PlainTextParam}(object, param, ...)

\S4method{readMsObject}{Spectra,PlainTextParam}(object, param, ...)

\S4method{saveObject}{Spectra}(x, path, ...)

\S4method{saveMsObject}{Spectra,AlabasterParam}(object, param, ...)

\S4method{readMsObject}{Spectra,AlabasterParam}(object, param, ...)
}
\arguments{
\item{object}{A \code{Spectra} object.}

\item{param}{Either a \code{PlainTextParam} or \code{AlabasterParam}.}

\item{...}{additional arguments passed to the \code{saveMsObject} or
\code{readMsObject} method of the \code{Spectra}'s \code{MsBackend}, such as
for example \code{consolidate} or \code{spectraPath}. See the \code{saveMsObject()}
and \code{readMsObject()} documentation of the used \code{MsBackend} class for
information on supported arguments.}

\item{x}{A \code{Spectra} object.}

\item{path}{For \code{saveObject()}: \code{character(1)} with the path where the
object should be stored in.}
}
\value{
\code{readMsObject()} returns a \link[Spectra:Spectra]{Spectra::Spectra} object.
}
\description{
\code{Spectra} objects can be stashed to (or read from) plain text file-based or
\emph{alabaster}-based formats using the \code{saveMsObject()} or \code{readMsObject()}
functions configured with the \link[MsStash:PlainTextParam]{PlainTextParam} or \link[MsStash:AlabasterParam]{AlabasterParam} parameter
objects, respectively. In both cases, the data from the \code{Spectra} object is
stored into the \emph{stash} directory defined with the parameter object.
Depending on the used MS backend (and parameters used), the stash can also
contain the MS data files.

At present, \code{Spectra} objects using one of the following MS data backends
are supported:
\itemize{
\item \code{MsBackendMzR}: see \link{MsBackendMzRStash} for details and options.
\item \code{MsBackendHdf5Peaks}: see \link{MsBackendHdf5PeaksStash} for details and
options.
}

The data backend of any \code{Spectra} backend can eventually be changed to one
of the above backends using the \code{\link[Spectra:setBackend]{Spectra::setBackend()}} method to support
saving the object into a \code{Spectra} stash. Support for additional backends
respectively data representations might also be provided by separate R
packages.

Details on the stash formats are provided in the respective sections below.
}
\section{\emph{alabaster}-based format, \code{AlabasterParam}}{


With \code{AlabasterParam}, the \code{Spectra} object will be exported (or imported)
through the \emph{alabaster} framework as a set of JSON and/or HDF5 files.
The content of each slot is stored to a separate file with the name
matching the slot name (converted to \emph{snake_case}). The object's \code{MsBackend}
is stored into a sub-folder \emph{backend} within the stash folder.

For information on the MS backend's data stash see the respective
documentation.
}

\section{Text-file format, \code{PlainTextParam}}{


For this format, the data content of a \code{Spectra} object is stored into the
files:
\itemize{
\item \emph{spectra_slots.txt}: plain text file containning the \emph{processing queue}
variables (separated by a \code{"|"}), the processing log messages (separated
by a \code{"|"}), the processing chunk size and the \code{MsBackend} class used.
\item \emph{spectra_processing_queue.json}: the object's processing queue serialized
in JSON format. It can be unserialized using \code{jsonlite::unserializeJSON()}.
}

For information on the MS backend's data see the respective documentation.
}

\examples{

## Create a Spectra object from two example MS data files (from MsDataHub)
library(Spectra)
library(MsDataHub)
s <- Spectra(
    c(X20171016_POOL_POS_1_105.134.mzML(),
      X20171016_POOL_POS_3_105.134.mzML()))
s

## Filter the intensities of the Spectra removing peaks with an intensity
## below 100
s <- filterIntensity(s, intensity = 100)

## Define the format and location of the `Spectra` stash: use the
## *alabaster*-based format and store the stash in a folder named
## *spectra_stash* in a temporary directory
d <- file.path(tempdir(), "spectra_stash")
ap <- AlabasterParam(d)

## Stash the `Spectra` object copying in addition the MS data files into the
## stash (`consolidate = TRUE`).
saveMsObject(s, ap, consolidate = TRUE)

## Show the content of the stash
dir(d)

## Read the `Spectra` object from the stash:
res <- readMsObject(Spectra(), ap)
res

## It is also possible to read individual contents from the stash. The
## directory *backend* contains for example the stashed `MsBackend` of the
## `Spectra` object. To read only the `MsBackend`:
ap2 <- AlabasterParam(file.path(d, "backend"))
b <- readMsObject(MsBackendMzR(), ap2)
b

## Alternatively, that data can also be read directly with the `readObject()`
## method from the *alabaster.base* package:
library(alabaster.base)
b <- readObject(file.path(d, "backend"))
b
}
\author{
Johannes Rainer, Philippine Louail
}
