% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsBackendCached.R
\name{MsBackendCachedStash}
\alias{MsBackendCachedStash}
\alias{saveMsObject,MsBackendCached,PlainTextParam-method}
\alias{readMsObject,MsBackendCached,PlainTextParam-method}
\alias{saveObject,MsBackendCached-method}
\alias{saveMsObject,MsBackendCached,AlabasterParam-method}
\alias{readMsObject,MsBackendCached,AlabasterParam-method}
\title{Stash for \code{MsBackendCached}}
\usage{
\S4method{saveMsObject}{MsBackendCached,PlainTextParam}(object, param, ...)

\S4method{readMsObject}{MsBackendCached,PlainTextParam}(object, param, ...)

\S4method{saveObject}{MsBackendCached}(x, path, ...)

\S4method{saveMsObject}{MsBackendCached,AlabasterParam}(object, param, ...)

\S4method{readMsObject}{MsBackendCached,AlabasterParam}(object, param, ...)
}
\arguments{
\item{object}{An \code{MsBackendCached} object.}

\item{param}{Either a \code{PlainTextParam} or \code{AlabasterParam}.}

\item{...}{Currently ignored.}

\item{x}{An \code{MsBackendCached} object.}

\item{path}{For \code{saveObject()}: \code{character(1)} with the path where the
object should be stored into.}
}
\value{
\code{readMsObject()} returns a \link[Spectra:MsBackendCached]{Spectra::MsBackendCached} object.
}
\description{
The \link[Spectra:MsBackendCached]{Spectra::MsBackendCached} backend keeps a cache of spectra metadata
(spectra variables) in an internal \code{data.frame} adding thus support for
(temporarily) changing or adding spectra variables for purely read-only
\code{MsBackend} classes. This backend is used e.g. by the
\link[MsBackendSql:MsBackendSql]{MsBackendSql::MsBackendSql} backend allowing to add or replace spectra
variables without affecting the content of the underlying data base used by
the \code{MsBackendSql}.

The stash of a \code{MsBackendCached} contains therefore the local spectra data
cache (if present), the names of the available spectra variables and the
total number of spectra. Supported stash formats are listed in the sections
below.
}
\details{
Notes for stash-functionality for \code{MsBackend} objects extending
\code{MsBackendCached}:
\itemize{
\item \code{saveMsObject()} and \code{saveObject()} will fail if the stash directory
already exist. Thus, stash functions of backend implementations extending
\code{MsBackendCached} should \strong{first} call the \code{MsBackendCached}'s
\code{saveMsObject()} or \code{saveObject()} \strong{before} exporting their respective
content to the stash directory.
}
}
\section{\emph{alabaster}-based format, \code{AlabasterParam}}{


The content from all slots of the \code{MsBackendCached} are stored using
functionality from the \emph{alabaster.base} package into separate sub-folders
of the base stash directory. These are:
\itemize{
\item \emph{local_data}: for the \code{data.frame} with the locally cached spectra
variables (slot \verb{@localData}).
\item \emph{nspectra}: (\code{integer(1)}) with the number of spectra.
\item \emph{spectra_variables}: (\code{character}) with the names of the object's spectra
variables.
}
}

\section{Text-file format, \code{PlainTextParam}}{


The data files written into the stash are:
\itemize{
\item \emph{ms_backend_data.txt}: tabulator separated text file with the content of
the \verb{@localData} slot.
\item \emph{ms_backend_nspectra.txt}: the number of spectra.
\item \emph{ms_backend_spectra_variables.txt}: the names of the object's spectra
variables (tabulator separated).
}
}

\examples{

library(Spectra)

## Create an empty `MsBackendCached` object
be <- MsBackendCached()

## Stash the object in alabaster-format in a temporary directory
ap <- AlabasterParam(file.path(tempdir(), "cache-stash"))
saveMsObject(be, ap)

## The content of the stash folder:
dir(file.path(tempdir(), "cache-stash"))

## Restore the object
res <- readMsObject(MsBackendCached(), ap)
res
}
\author{
Johannes Rainer
}
