% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsBackendMemory.R
\name{MsBackendMemoryStash}
\alias{MsBackendMemoryStash}
\alias{saveMsObject,MsBackendMemory,PlainTextParam-method}
\alias{readMsObject,MsBackendMemory,PlainTextParam-method}
\alias{saveObject,MsBackendMemory-method}
\alias{saveMsObject,MsBackendMemory,AlabasterParam-method}
\title{Stash for \code{MsBackendMemory}}
\usage{
\S4method{saveMsObject}{MsBackendMemory,PlainTextParam}(object, param, ...)

\S4method{readMsObject}{MsBackendMemory,PlainTextParam}(object, param, ...)

\S4method{saveObject}{MsBackendMemory}(x, path, ...)

\S4method{saveMsObject}{MsBackendMemory,AlabasterParam}(object, param, ...)
}
\arguments{
\item{object}{An \code{MsBackendMemory} object.}

\item{param}{Either a \code{PlainTextParam} or \code{AlabasterParam}.}

\item{...}{Currently ignored.}

\item{x}{An \code{MsBackendMemory} object.}

\item{path}{For \code{saveObject()}: \code{character(1)} with the path where the
object should be stored into.}
}
\description{
The stash for \link[Spectra:MsBackendMemory]{Spectra::MsBackendMemory} contains both the spectra variables
(spectra metadata) as well as the peaks data (\emph{m/z} and intensity values).
}
\details{
To support storing spectra \strong{and} peaks variables, the \code{MsBackendMemory} is
converted to a \link[Spectra:MsBackendHdf5Peaks]{Spectra::MsBackendHdf5Peaks} object with the full peaks data
stored into a \emph{peaks.h5} HDF5 file within the stash directory. Due to the
nested directory structure it is possible to load a \code{MsBackendMemory} stash
either as a \code{MsBackendMemory} or a \code{MsBackendHdf5Peaks}.
}
\note{
Currently only peaks variables \code{"mz"} and \code{"intensity"} are supported.
}
\section{Text-file format, \code{PlainTextParam}}{


The \code{MsBackendMemory} stash contains a text-file format
\link{MsBackendHdf5PeaksStash} within a sub-folder \emph{backend}. When loading,
the data is first imported as a \code{MsBackendHdf5Peaks} which is then converted
to the resulting \code{MsBackendMemory} (with all data materialized into memory).
}

\section{\emph{alabaster}-based format, \code{AlabasterParam}}{


The full data is stored as an \emph{alabaster}-format \link{MsBackendHdf5PeaksStash}
in a directory \emph{backend} within the stash folder. Due to this nesting it is
possible to load the data either as \code{MsBackendMemory} or
\code{MsBackendHdf5Peaks}.
}

\examples{

library(Spectra)

}
\author{
Johannes Rainer
}
