% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsBackendMzR.R
\name{MsBackendMzR-stash}
\alias{MsBackendMzR-stash}
\alias{saveMsObject,MsBackendMzR,PlainTextParam-method}
\alias{readMsObject,MsBackendMzR,PlainTextParam-method}
\alias{saveObject,MsBackendMzR-method}
\title{MsBackendMzR Stash}
\usage{
\S4method{saveMsObject}{MsBackendMzR,PlainTextParam}(object, param)

\S4method{readMsObject}{MsBackendMzR,PlainTextParam}(object, param, spectraPath = character())

\S4method{saveObject}{MsBackendMzR}(x, path, ...)
}
\arguments{
\item{object}{An \code{MsBackendMzR} object.}

\item{param}{Either a \code{PlainTextParam} or \code{AlabasterParam}.}

\item{spectraPath}{For \code{readMsObject()}: optional \code{character(1)} with the
path to the MS data files (mzML, mzXML or CDF) in case they are on longer
available in the folder referred to by the original stashed
\code{MsBackendMzR} object.}

\item{x}{An \code{MsBackendMzR} object.}

\item{path}{For \code{saveObject()}: \code{character(1)} with the path where the
object should be stored in.}
}
\value{
\code{readMsObject()} returns an \link[Spectra:MsBackend]{Spectra::MsBackendMzR}` object.
}
\description{
\code{MsBackendMzR} classes can be stashed to (or read from) plain text
file-based or \emph{alabaster}-based formats using the \code{\link[MsStash:saveMsObject]{MsStash::saveMsObject()}} and
\code{\link[MsStash:saveMsObject]{MsStash::readMsObject()}} functions combined with the \link[MsStash:PlainTextParam]{MsStash::PlainTextParam} and
\link[MsStash:AlabasterParam]{MsStash::AlabasterParam} parameter objects, respectively. The properties for both
formats are described in detail in the sections below.
}
\section{Text-file format, \code{PlainTextParam}}{


The \code{saveMsObject()} function with the \code{PlainTextParam} stores the spectra
metadata (spectra variables) of an \code{MsBackendMzR} to a plain tabulator
delimited text file with the name \emph{ms_backend_spectra_data.txt} in the
directory specified with parameter \code{path} of the \code{PlainTextParam} object.
Importantly, the peaks data (the \emph{m/z} and intensity values) are \strong{not}
exported with \code{saveMsObject()}. \code{readMsObject()} restores a previously
stashed \code{MsBackendMzR} object from the directory specified with parameter
\code{path} of the \code{PlainTextParam}.

The additional parameter \code{spectraPath} of \code{readMsObject()} allows to define
the path to the MS data files containing the full MS data (i.e., the mzML,
mzXML or CDF files referred to by the \code{MsBackendMzR}).
}

\section{\emph{alabaster}-based format, \code{AlabasterParam}}{

}

\author{
Philippine Louail, Johannes Rainer
}
